+++ .circleci/config.yml
@@ -21,7 +21,7 @@ references:
      name: Testing
      command: |
        python --version ; pip --version ; pip list
-       py.test pytorch_lightning tests -v --junitxml=test-reports/pytest_junit.xml
+       py.test pytorch_lightning tests pl_develop -v --junitxml=test-reports/pytest_junit.xml
      no_output_timeout: 15m
 
   examples: &examples
