+++ benchmarks/test_parity.py
@@ -4,9 +4,9 @@ import numpy as np
 import pytest
 import torch
 
-import tests.base.utils as tutils
+import pl_develop.utils as tutils
 from pytorch_lightning import Trainer, seed_everything
-from tests.base.models import ParityModuleRNN, ParityModuleMNIST
+from pl_develop import ParityModuleRNN, ParityModuleMNIST
 
 
 @pytest.mark.parametrize('cls_model,max_diff', [
