+++ pl_develop/__init__.py
@@ -1,0 +1,5 @@
+"""Models for testing."""
+
+from pl_develop.model_template import EvalModelTemplate
+from pl_develop.datasets import TrialMNIST, TrialCIFAR10
+from pl_develop.models import EvalModelGAN, ParityModuleMNIST, ParityModuleRNN
