--- tests/base/__init__.py
@@ -1,3 +1,0 @@
-"""Models for testing."""
-
-from tests.base.model_template import EvalModelTemplate
