+++ tests/callbacks/test_callbacks.py
@@ -8,7 +8,7 @@ from pytorch_lightning import Callback
 from pytorch_lightning import Trainer, LightningModule
 from pytorch_lightning.callbacks import EarlyStopping, ModelCheckpoint
 from pytorch_lightning.loggers import TensorBoardLogger
-from tests.base import EvalModelTemplate
+from pl_develop import EvalModelTemplate
 
 
 def test_early_stopping_functionality(tmpdir):
