+++ tests/callbacks/test_progress_bar.py
@@ -3,7 +3,7 @@ import pytest
 from pytorch_lightning import Trainer
 from pytorch_lightning.callbacks import ProgressBarBase, ProgressBar, ModelCheckpoint
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
-from tests.base import EvalModelTemplate
+from pl_develop import EvalModelTemplate
 
 
 @pytest.mark.parametrize('callbacks,refresh_rate', [
