+++ tests/models/data/horovod/train_default_model.py
@@ -30,9 +30,9 @@ sys.path.insert(0, os.path.abspath(PATH_ROOT))
 
 from pytorch_lightning import Trainer  # noqa: E402
 from pytorch_lightning.callbacks import ModelCheckpoint  # noqa: E402
-from tests.base import EvalModelTemplate  # noqa: E402
-from tests.base.pipelines import run_model_test  # noqa: E402
-from tests.base.utils import set_random_master_port  # noqa: E402
+from pl_develop import EvalModelTemplate  # noqa: E402
+from pl_develop import run_model_test  # noqa: E402
+from pl_develop import set_random_master_port  # noqa: E402
 
 
 parser = argparse.ArgumentParser()
