+++ tests/models/test_cpu.py
@@ -10,7 +10,7 @@ import tests.base.utils as tutils
 from pytorch_lightning import Trainer
 from pytorch_lightning.callbacks import EarlyStopping
 from pytorch_lightning.callbacks import ModelCheckpoint
-from tests.base import EvalModelTemplate
+from pl_develop import EvalModelTemplate
 
 
 def test_cpu_slurm_save_load(tmpdir):
