+++ tests/test_deprecated.py
@@ -6,7 +6,7 @@ import pytest
 import torch
 
 from pytorch_lightning import Trainer
-from tests.base import EvalModelTemplate
+from pl_develop import EvalModelTemplate
 
 
 def _soft_unimport_module(str_module):
