+++ tests/trainer/test_lr_finder.py
@@ -3,7 +3,7 @@ import torch
 
 from pytorch_lightning import Trainer
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
-from tests.base import EvalModelTemplate
+from pl_develop import EvalModelTemplate
 
 
 def test_error_on_more_than_1_optimizer(tmpdir):
